requirements.txt
autocorrect==0.3.0
certifi==2017.4.17
chardet==3.0.4
idna==2.5
nltk==3.2.4
numpy==1.13.0

requests==2.17.3
six==1.10.0
urllib3==1.21.1
websocket-client==0.42.1
